auto selector (does not work due to lifetimes)
