+++ Cargo.toml
@@ -17,6 +17,9 @@ futures-lite = "2.2.0"
 tokio = { version = "1.36.0", features = ["rt", "full"] }
 serde_json = "1.0.113"
 serde = "1.0.196"
+lazy_static = "1.4.0"
+send_wrapper = "0.6.0"
+once_cell = "1.19.0"
 
 [dependencies.frcrs]
 git = "https://www.github.com/Team-2502/frcrs.git"
